+++ test/integration/test_keyprotect.py
@@ -1,7 +1,6 @@
 import logging
 import time
 import unittest
-from datetime import datetime, timedelta
 
 import redstone
 
@@ -367,7 +366,7 @@ class KeyProtectTestCase(unittest.TestCase):
 
         cert_created = self.kp.kmip_cert_create(
             adapter_id,
-            generate_selfsigned_cert("www.kms.test.cloud.ibm.com"),
+            self_signed_cert.generate_selfsigned_cert("www.kms.test.cloud.ibm.com"),
             adapter_id,
             "mycert",
         )
@@ -379,54 +378,6 @@ class KeyProtectTestCase(unittest.TestCase):
 
         self.kp.kmip_cert_delete(adapter_id, cert_created.get("id"))
         self.kp.kmip_adapter_delete(adapter_id)
-
-
-### Utility Methods
-
-
-# Utility method to generate a certificate for the purposes of testing the KMIP mgmt apis
-def generate_selfsigned_cert(hostname, key=None):
-    """Generates self signed certificate for a hostname, and optional IP addresses."""
-    from cryptography import x509
-    from cryptography.x509.oid import NameOID
-    from cryptography.hazmat.primitives import hashes
-    from cryptography.hazmat.backends import default_backend
-    from cryptography.hazmat.primitives import serialization
-    from cryptography.hazmat.primitives.asymmetric import rsa
-
-    # Generate our key
-    if key is None:
-        key = rsa.generate_private_key(
-            public_exponent=65537,
-            key_size=2048,
-            backend=default_backend(),
-        )
-
-    name = x509.Name([x509.NameAttribute(NameOID.COMMON_NAME, hostname)])
-
-    # best practice seem to be to include the hostname in the SAN, which *SHOULD* mean COMMON_NAME is ignored.
-    alt_names = [x509.DNSName(hostname)]
-
-    san = x509.SubjectAlternativeName(alt_names)
-
-    # path_len=0 means this cert can only sign itself, not other certs.
-    basic_contraints = x509.BasicConstraints(ca=True, path_length=0)
-    now = datetime.now(datetime.UTC)
-    cert = (
-        x509.CertificateBuilder()
-        .subject_name(name)
-        .issuer_name(name)
-        .public_key(key.public_key())
-        .serial_number(1000)
-        .not_valid_before(now)
-        .not_valid_after(now + timedelta(days=10 * 365))
-        .add_extension(basic_contraints, False)
-        .add_extension(san, False)
-        .sign(key, hashes.SHA256(), default_backend())
-    )
-    cert_pem = cert.public_bytes(encoding=serialization.Encoding.PEM)
-
-    return cert_pem.decode("utf-8")
 
 
 if __name__ == "__main__":
